Add files via upload
Simple Blog Display
This project is a simple blog display website that fetches and displays a list of blog posts from a static JSON file. The primary aim of this project is to introduce working with external data in a web development context.

Features
Header with Navigation Menu: The website has a header section with a logo on the left and navigation links to different sections (Blogs, About, Contact) on the right.
Full-Screen Background Image: The homepage features a full-screen background image that spans the entire viewport height.
About Section: Provides information about the website, linked in the navigation menu.
Blogs Section: Displays blog posts fetched from a static JSON file, specifically focusing on skincare topics.
Contact Section: Contains contact information, linked in the navigation menu.
Footer: Includes a footer with a project credit.